+++ service/conversations/v1/service/api_op_client.go
@@ -6,6 +6,8 @@ import (
 	"github.com/RJPearson94/twilio-sdk-go/service/conversations/v1/service/binding"
 	"github.com/RJPearson94/twilio-sdk-go/service/conversations/v1/service/bindings"
 	"github.com/RJPearson94/twilio-sdk-go/service/conversations/v1/service/configuration"
+	"github.com/RJPearson94/twilio-sdk-go/service/conversations/v1/service/conversation"
+	"github.com/RJPearson94/twilio-sdk-go/service/conversations/v1/service/conversations"
 	"github.com/RJPearson94/twilio-sdk-go/service/conversations/v1/service/role"
 	"github.com/RJPearson94/twilio-sdk-go/service/conversations/v1/service/roles"
 	"github.com/RJPearson94/twilio-sdk-go/service/conversations/v1/service/user"
@@ -22,6 +24,8 @@ type Client struct {
 	Binding       func(string) *binding.Client
 	Bindings      *bindings.Client
 	Configuration func() *configuration.Client
+	Conversation  func(string) *conversation.Client
+	Conversations *conversations.Client
 	Role          func(string) *role.Client
 	Roles         *roles.Client
 	User          func(string) *user.Client
@@ -54,6 +58,15 @@ func New(client *client.Client, properties ClientProperties) *Client {
 				ServiceSid: properties.Sid,
 			})
 		},
+		Conversation: func(conversationSid string) *conversation.Client {
+			return conversation.New(client, conversation.ClientProperties{
+				ServiceSid: properties.Sid,
+				Sid:        conversationSid,
+			})
+		},
+		Conversations: conversations.New(client, conversations.ClientProperties{
+			ServiceSid: properties.Sid,
+		}),
 		Role: func(roleSid string) *role.Client {
 			return role.New(client, role.ClientProperties{
 				ServiceSid: properties.Sid,